// studio/datart-versions/datart-1.0.0-rc.1/frontend/src/app/types/ChartConfig.ts
export type ChartConfigBase = {
  label?: string;
  key: string;
};
// ...
export type ChartStyleSectionTemplate = {
  template?: ChartStyleSectionRow;
};

export type ChartStyleConfig = ChartConfigBase &
  ChartStyleSectionGroup &
  ChartStyleSectionTemplate;

export type ChartStyleSectionGroup = ChartStyleSectionRow & {
  rows?: ChartStyleSectionGroup[];
};

export type ChartStyleSectionRow = {
  label: string;
  key: string;
  default?: any;
  value?: any;
  disabled?: boolean;
  hide?: boolean;
  options?: ChartStyleSectionRowOption;
  watcher?: ChartStyleSectionRowWatcher;
  template?: ChartStyleSectionRow;
  comType: ValueOf<typeof ChartStyleSectionComponentType>;
  hidden?: boolean;
};